Create initial SurveyUser after login.
Aparantly, this was always done on the fly. Previously in the function select_user

For the far future: it's probably better to create the SurveyUser after register.
Although: given the fact that survey users may be deleted at random moments,
having guards in the locations where they're actually required isn't so crazy
either.