update_data() should *not* swallow exceptions
in our infrastructure in particular (psql) this will only
lead to hard to debug errors like the below (in the next statement).
better to raise the error when it happens.

update_data() is only called in the chart_update() command anyway,
so no further consequences are predicted.

django.db.utils.DatabaseError: current transaction is aborted, commands ignored until end of transaction block